requirements.txt
aiocron==1.8
aiofiles==0.7.0
aiohttp==3.7.4.post0
Async-OAuthlib==0.0.9
async-timeout==3.0.1
asyncio==3.4.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
discord.py==1.7.3
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2